stg-synthetic-eval/requirements.txt
numpy
pandas
matplotlib
tqdm
torch
pyyaml
scipy
